+++ shared/include/srvros/syscall_numbers.h
@@ -64,6 +64,8 @@
 #define SYS_FCHMOD 59
 #define SYS_STATFS 60
 #define SYS_PROC_GROUP 61
+#define SYS_MMAP 62
+#define SYS_MUNMAP 63
 
 #define SRV_OPEN_READ 0x01
 #define SRV_OPEN_WRITE 0x02
@@ -146,6 +148,16 @@ struct srv_winsize {
 
 #define SRV_EXEC_BACKGROUND 0x01
 #define SRV_EXEC_REPLACE 0x02
+
+#define SRV_PROT_NONE 0x00
+#define SRV_PROT_READ 0x01
+#define SRV_PROT_WRITE 0x02
+#define SRV_PROT_EXEC 0x04
+
+#define SRV_MAP_SHARED 0x01
+#define SRV_MAP_PRIVATE 0x02
+#define SRV_MAP_FIXED 0x10
+#define SRV_MAP_ANONYMOUS 0x20
 
 #define SRV_POLLIN 0x0001
 #define SRV_POLLOUT 0x0004